Refactored performance and points in to a trait as they are shared by Team and Manager
