+++ app/Traits/HasFantasyPoints.php
@@ -1,0 +1,19 @@
+<?php namespace Fantasee\Traits;
+
+trait HasFantasyPoints {
+
+  /**
+  * Get points attributes
+  *
+  * @var number
+  */
+  public function getPointsAttribute()
+  {
+    $object = new \StdClass;
+
+    $object->for = $this->getPointsFor();
+    $object->against = $this->getPointsAgainst();
+
+    return $object;
+  }
+}
