+++ app/Traits/HasPerformanceRecord.php
@@ -1,0 +1,17 @@
+<?php namespace Fantasee\Traits;
+
+trait HasPerformanceRecord {
+  /**
+  * Get hash of team statistics
+  *
+  * @var number
+  */
+  public function getPerformanceAttribute()
+  {
+    return [
+      'wins'   => $this->wins,
+      'losses' => $this->losses,
+      'ties'   => $this->ties,
+    ];
+  }
+}
